Win/loss conditions (small change)
Added win/loss conditions to NextDayButton_Click method. This way, after
each day passes, we check that there are no more travelers (loss) or the
player gets past day 10 (win).

Did not write what to do in these conditions, but it could be either a
separate window or a message on the game screen declaring a victory or
loss. Up to you guys.